Void functions are now int functions
Handling any error message will be easier after that change -- it will
be even possible to integrate HDF error reporting to Mechanic error
reporting (which is still to be build). Modules can now returns integer
value to show if everything works ok or not.

Some documentation enhancements in the default module.