MPI as a hard dependency
I decided to leave MPI as a hard requirement to run Mechanic, since even in single-cpu
environments it is easy to simulate MPI nodes (I called it fake-MPI) -- I tested it on a
x86 machine and it worked quite well. Running Mechanic only on one node will cause to use
Masteralone mode automatically, despite of any commandline or config options. And it will
be much easier to handle for the build system.