Master sends initial conditions to slaves
There is an initial solution for bug #9 (Master node does not send initial
condition). There is now node_preparePixel() function provided and node_in()
can also setup initial condition (depends on design of the module). The lenght
of initial condition can be adjusted with md->irl in module_init(). The initial
condition has the same type as result struct, and is sended to each slave at
each pixel, so it is possible to modify the initials during computations.
Depending on setup, the master node can read config file or each slave own.
There is of course need for improvements and clean-up.