Solved #16, segmentation fault on checkpoints
This specific error appears when you have number of outstanding receives
greater than size of the checkpoint, and the loop tries to save data outside
the array. Just a little change to counters and additional check for the
checkpoint in outstanding results shouls solve this problem. The Masteralone
mode is not affected.