Fixed issue #21: Restart problem with first checkpoint file
The problem was with backup stage of the master file -- the rename function was changing
the name of the master file and thus, the default mechanic-master-00.h5 was not present.
Then, the code tried to read the master file, and, because it was not present, the error
occured. The problem was only when restarting with the first checkpoint (the master file).
This commit should fix this issue.