F2003 full module example
Fortran2003 full example module has been added, as well as some fixes to the build system.
We switch to -j1 (one process) during compilation, just because this is the only way to
compile and link fortran modules properly (not the fortran bindings itself, you could
build them without any problems). There is a small addition to moduleInfo struct -- api
info, for future development, and possible changes in the module api. By default, its
value is set to 1. Modules are unaffected with this change.