Restart mode: functional implementation
We have now functional implementation of the restart mode. The setup is properly read from
the restart file, as well as data stored in all datasets (note: STORAGE_BASIC for tasks is
not implemented yet). The restart will be performed from the last pool, if any of tasks
are available.