Initial implementation of the new memory allocator
The new allocator will allow to use datasets of different datatypes and ranks. New
functions are available in API: SetData() which would be use to put current data to
the core memory and GetData() which would be use to get the data from memory buffers.
These functions will be used instead of direct t->storage[0].data[0][0] etc.