Initial commit for 2.2.x
Datasets of different datatypes basically work. The board is not ported yet, as well as
Non-blocking communication mode. All MPI/HDF5 basic datatypes are supported. New functions
for getting and setting data are available, SetData() and GetData(), which must be used for
proper data handling. The double data[][] array is deprecated by now. The core module
brings some basic tests for further development, which will be removed in the final
version.